Replace Travis CI with Github Actions
Since travis-ci.org will be shutdown on June 15th, 2021 we need to move
to an alternative CI service.  This PR replaces Travis with GitHub
Actions, which is the obvious choice for a GitHub repository.
Note that the CI works as before, meaning that every push to `content`
will trigger deployment of the changed content.

Deployment on `master` is done with this GitHub Action:
https://github.com/JamesIves/github-pages-deploy-action